Improve types. Fix compile error
